redux audit
