+++ gui/src/App.tsx
@@ -1,11 +1,8 @@
 import { useDispatch } from "react-redux";
 import { RouterProvider, createMemoryRouter } from "react-router-dom";
 import Layout from "./components/Layout";
-import { SubmenuContextProvidersContext } from "./context/SubmenuContextProviders";
-import { VscThemeContext } from "./context/VscTheme";
+import { VscThemeProvider } from "./context/VscTheme";
 import useSetup from "./hooks/useSetup";
-import useSubmenuContextProviders from "./hooks/useSubmenuContextProviders";
-import { useVscTheme } from "./hooks/useVscTheme";
 import { AddNewModel, ConfigureProvider } from "./pages/AddNewModel";
 import ConfigErrorPage from "./pages/config-error";
 import Edit from "./pages/Edit";
@@ -18,6 +15,7 @@ import MorePage from "./pages/More";
 import SettingsPage from "./pages/settings";
 import Stats from "./pages/stats";
 import { ROUTES } from "./util/navigation";
+import { SubmenuContextProvidersProvider } from "./context/SubmenuContextProviders";
 
 const router = createMemoryRouter([
   {
@@ -77,22 +75,25 @@ const router = createMemoryRouter([
   },
 ]);
 
-function App() {
+/*
+  Prevents entire app from rerendering continuously with useSetup in App
+  TODO - look into a more redux-esque way to do this
+*/
+function SetupListeners() {
   const dispatch = useDispatch();
 
   useSetup(dispatch);
+  return <></>;
+}
 
-  const vscTheme = useVscTheme();
-  const submenuContextProvidersMethods = useSubmenuContextProviders();
-
+function App() {
   return (
-    <VscThemeContext.Provider value={vscTheme}>
-      <SubmenuContextProvidersContext.Provider
-        value={submenuContextProvidersMethods}
-      >
+    <VscThemeProvider>
+      <SubmenuContextProvidersProvider>
         <RouterProvider router={router} />
-      </SubmenuContextProvidersContext.Provider>
-    </VscThemeContext.Provider>
+      </SubmenuContextProvidersProvider>
+      <SetupListeners />
+    </VscThemeProvider>
   );
 }
