+++ gui/src/components/markdown/StepContainerPreToolbar/StepContainerPreToolbar.tsx
@@ -1,6 +1,6 @@
 import { ChevronDownIcon } from "@heroicons/react/24/outline";
 import { debounce } from "lodash";
-import { useContext, useEffect, useRef, useState } from "react";
+import { useContext, useEffect, useMemo, useRef, useState } from "react";
 import { useSelector } from "react-redux";
 import styled, { css, keyframes } from "styled-components";
 import { v4 as uuidv4 } from "uuid";
@@ -80,11 +80,12 @@ export default function StepContainerPreToolbar(
     (state: RootState) => state.state.nextCodeBlockToApplyIndex,
   );
 
-  const applyState = useSelector((store: RootState) =>
-    store.state.applyStates.find(
-      (state) => state.streamId === streamIdRef.current,
-    ),
+  const applyStates = useSelector(
+    (store: RootState) => store.state.applyStates,
   );
+  const applyState = useMemo(() => {
+    return applyStates.find((state) => state.streamId === streamIdRef.current);
+  }, [applyStates, streamIdRef]);
 
   // This handles an edge case when the last node in the markdown syntax tree is a codeblock.
   // In this scenario, `isGeneratingCodeBlock` is never set to false since we determine if
