--- gui/src/context/VscTheme.ts
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const VscThemeContext = createContext<any>(undefined);
