+++ gui/src/editorInset/EditorInset.tsx
@@ -5,7 +5,7 @@ import { defaultBorderRadius } from "../components";
 import TipTapEditor from "../components/mainInput/TipTapEditor";
 import { IdeMessengerContext } from "../context/IdeMessenger";
 import useSetup from "../hooks/useSetup";
-import { selectSlashCommands } from "../redux/selectors";
+import { selectSlashCommandComboBoxInputs } from "../redux/selectors";
 import { RootState } from "../redux/store";
 
 const EditorInsetDiv = styled.div`
@@ -18,7 +18,7 @@ const EditorInsetDiv = styled.div`
 
 function EditorInset() {
   const dispatch = useDispatch();
-  const availableSlashCommands = useSelector(selectSlashCommands);
+  const availableSlashCommands = useSelector(selectSlashCommandComboBoxInputs);
   const availableContextProviders = useSelector(
     (store: RootState) => store.state.config.contextProviders,
   );
