+++ gui/src/pages/gui/Chat.tsx
@@ -12,6 +12,7 @@ import {
   useCallback,
   useContext,
   useEffect,
+  useMemo,
   useRef,
   useState,
 } from "react";
@@ -131,9 +132,14 @@ export function Chat() {
   const mainTextInputRef = useRef<HTMLInputElement>(null);
   const stepsDivRef = useRef<HTMLDivElement>(null);
   const [isAtBottom, setIsAtBottom] = useState<boolean>(false);
-  const state = useSelector((state: RootState) => state.state);
+  const history = useSelector((store: RootState) => store.state.history);
+  const showChatScrollbar = useSelector(
+    (store: RootState) => store.state.config?.ui?.showChatScrollbar,
+  );
   const { saveSession, getLastSessionId, loadLastSession } =
     useHistory(dispatch);
+
+  console.log("chat rendering");
 
   const snapToBottom = useCallback(() => {
     if (!stepsDivRef.current) return;
@@ -289,7 +295,7 @@ export function Chat() {
         setLocalStorage("mainTextEntryCounter", 1);
       }
     },
-    [state.history, defaultModel, state, streamResponse],
+    [history, defaultModel, streamResponse],
   );
 
   useWebviewListener(
@@ -303,26 +309,27 @@ export function Chat() {
 
   const isLastUserInput = useCallback(
     (index: number): boolean => {
-      return !state.history
+      return !history
         .slice(index + 1)
         .some((entry) => entry.message.role === "user");
     },
-    [state.history],
+    [history],
   );
 
-  const showScrollbar =
-    state.config.ui?.showChatScrollbar || window.innerHeight > 5000;
+  const showScrollbar = useMemo(() => {
+    return showChatScrollbar ?? window.innerHeight > 5000;
+  }, [showChatScrollbar]);
 
   return (
     <>
       {widget}
       <StepsDiv
         ref={stepsDivRef}
-        className={`overflow-y-scroll pt-[8px] ${showScrollbar ? "thin-scrollbar" : "no-scrollbar"} ${state.history.length > 0 ? "flex-1" : ""}`}
+        className={`overflow-y-scroll pt-[8px] ${showScrollbar ? "thin-scrollbar" : "no-scrollbar"} ${history.length > 0 ? "flex-1" : ""}`}
         onScroll={handleScroll}
       >
         {highlights}
-        {state.history.map((item, index: number) => (
+        {history.map((item, index: number) => (
           <Fragment key={item.message.id}>
             <ErrorBoundary
               FallbackComponent={fallbackRender}
@@ -368,7 +375,7 @@ export function Chat() {
                   >
                     <StepContainer
                       index={index}
-                      isLast={index === state.history.length - 1}
+                      isLast={index === history.length - 1}
                       item={item}
                     />
                   </TimelineItem>
@@ -421,7 +428,7 @@ export function Chat() {
         >
           <div className="flex flex-row items-center justify-between pb-1 pl-0.5 pr-2">
             <div className="xs:inline hidden">
-              {state.history.length === 0 && getLastSessionId() ? (
+              {history.length === 0 && getLastSessionId() ? (
                 <div className="xs:inline hidden">
                   <NewSessionButton
                     onClick={async () => {
@@ -440,7 +447,7 @@ export function Chat() {
             <ConfigErrorIndicator />
           </div>
 
-          {state.history.length === 0 && (
+          {history.length === 0 && (
             <>
               {onboardingCard.show && (
                 <div className="mx-2 mt-10">
@@ -458,7 +465,7 @@ export function Chat() {
         </div>
       </div>
       <div
-        className={`${state.history.length === 0 ? "h-full" : ""} flex flex-col justify-end`}
+        className={`${history.length === 0 ? "h-full" : ""} flex flex-col justify-end`}
       >
         <ChatIndexingPeeks />
       </div>
