+++ gui/src/redux/selectors/modelSelectors.ts
@@ -1,11 +1,12 @@
-import { DEFAULT_MAX_TOKENS } from "core/llm/constants";
 import { RootState } from "../store";
+import { createSelector } from "@reduxjs/toolkit";
 
-export const defaultModelSelector = (state: RootState) => {
-  const title = state.state.defaultModelTitle;
-  return state.state.config.models.find((model) => model.title === title);
-};
-
-export const contextLengthSelector = (state: RootState) => {
-  return defaultModelSelector(state)?.contextLength || DEFAULT_MAX_TOKENS;
-};
+export const defaultModelSelector = createSelector(
+  [
+    (store: RootState) => store.state.defaultModelTitle,
+    (store: RootState) => store.state.config.models,
+  ],
+  (defaultModelTitle, models) => {
+    return models?.find((model) => model.title === defaultModelTitle);
+  },
+);
