+++ gui/src/redux/slices/stateSlice.ts
@@ -341,7 +341,7 @@ export const stateSlice = createSlice({
         payload.rangeInFileWithContents.range.start.line + 1
       }-${payload.rangeInFileWithContents.range.end.line + 1})`;
       contextItems.push({
-        name: `${base} ${lineNums}`,
+        name: `${base} ${lineNums}s`,
         description: payload.rangeInFileWithContents.filepath,
         id: {
           providerTitle: "code",
@@ -367,10 +367,7 @@ export const stateSlice = createSlice({
       };
     },
     setSelectedProfileId: (state, { payload }: PayloadAction<string>) => {
-      return {
-        ...state,
-        selectedProfileId: payload,
-      };
+      state.selectedProfileId = payload;
     },
     setCurCheckpointIndex: (state, { payload }: PayloadAction<number>) => {
       state.curCheckpointIndex = payload;
